Correct placement of placeholders

+++ packages/flutter/lib/src/services/text_input.dart
@@ -1164,13 +1164,13 @@ class TextInput {
     if (method == 'TextInputClient.focusElement') {
       final List<dynamic> args = methodCall.arguments as List<dynamic>;
       if (_scribbleClients.containsKey(args[0])) {
-        _scribbleClients[args[0]]?.onScribbleFocus(Offset(args[1].toDouble(), args[2].toDouble()));
+        _scribbleClients[args[0]]?.onScribbleFocus(Offset(args[1].toDouble() as double, args[2].toDouble() as double));
       }
       return;
     } else if (method == 'TextInputClient.requestElementsInRect') {
       final List<dynamic> args = methodCall.arguments as List<dynamic>;
       return _scribbleClients.keys.where((String elementIdentifier) {
-        final Rect rect = Rect.fromLTWH(args[0].toDouble() , args[1].toDouble() , args[2].toDouble() , args[3].toDouble() );
+        final Rect rect = Rect.fromLTWH(args[0].toDouble() as double, args[1].toDouble() as double, args[2].toDouble() as double, args[3].toDouble() as double);
         return _scribbleClients[elementIdentifier]?.isInScribbleRect(rect) ?? false;
       }).map((String elementIdentifier) {
         final Rect bounds = _scribbleClients[elementIdentifier]?.bounds ?? Rect.zero;
